@@ -1,10 +1,10 @@
 {
-  "name": "hello-redis",
+  "name": "redis-web-client",
   "version": "1.0.0",
   "scripts": {
     "ng": "ng",
     "start": "ng serve --open",
-    "build": "ng build --prod --aot",
+    "build": "ng build --configuration production --aot",
     "test": "ng test",
     "lint": "ng lint",
     "e2e": "ng e2e"
@@ -16,39 +16,39 @@
   },
   "license": "MIT",
   "dependencies": {
-    "@angular/animations": "~10.0.9",
-    "@angular/common": "~10.0.9",
-    "@angular/compiler": "~10.0.9",
-    "@angular/core": "~10.0.9",
-    "@angular/forms": "~10.0.9",
-    "@angular/platform-browser": "~10.0.9",
-    "@angular/platform-browser-dynamic": "~10.0.9",
-    "@angular/router": "~10.0.9",
-    "@angular/service-worker": "~10.0.9",
+    "@angular/animations": "~13.1.3",
+    "@angular/common": "~13.1.3",
+    "@angular/compiler": "~13.1.3",
+    "@angular/core": "~13.1.3",
+    "@angular/forms": "~13.1.3",
+    "@angular/platform-browser": "~13.1.3",
+    "@angular/platform-browser-dynamic": "~13.1.3",
+    "@angular/router": "~13.1.3",
+    "@angular/service-worker": "~13.1.3",
     "ng-zorro-antd": "13.0.1",
-    "rxjs": "~6.6.7",
+    "rxjs": "~7.5.2",
     "tslib": "^2.3.1",
-    "zone.js": "~0.10.3"
+    "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.1000.8",
-    "@angular/cli": "~10.0.6",
-    "@angular/compiler-cli": "~10.0.9",
+    "@angular-devkit/build-angular": "~13.1.4",
+    "@angular/cli": "~13.1.4",
+    "@angular/compiler-cli": "~13.1.3",
     "@angular/language-service": "~13.1.3",
-    "@types/jasmine": "~3.3.8",
-    "@types/jasminewd2": "~2.0.3",
-    "@types/node": "^17.0.8",
-    "codelyzer": "^6.0.0",
-    "jasmine-core": "~3.5.0",
+    "@types/jasmine": "~3.10.3",
+    "@types/jasminewd2": "~2.0.10",
+    "@types/node": "^17.0.10",
+    "codelyzer": "^6.0.2",
+    "jasmine-core": "~4.0.0",
     "jasmine-spec-reporter": "~7.0.0",
-    "karma": "~5.0.0",
+    "karma": "~6.3.11",
     "karma-chrome-launcher": "~3.1.0",
-    "karma-coverage-istanbul-reporter": "~3.0.2",
-    "karma-jasmine": "~3.3.0",
-    "karma-jasmine-html-reporter": "^1.5.0",
+    "karma-coverage-istanbul-reporter": "~3.0.3",
+    "karma-jasmine": "~4.0.1",
+    "karma-jasmine-html-reporter": "^1.7.0",
     "protractor": "~7.0.0",
-    "ts-node": "~7.0.0",
+    "ts-node": "~10.4.0",
     "tslint": "~6.1.0",
-    "typescript": "~3.9.10"
+    "typescript": "~4.5.5"
   }
 }
