@@ -1,17 +1,16 @@
 import { NgModule } from '@angular/core';
-import { HttpClientModule, HTTP_INTERCEPTORS } from '@angular/common/http';
+import { HttpClientModule } from '@angular/common/http';
 import { BrowserModule } from '@angular/platform-browser';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
-import { NZ_I18N, zh_CN, NZ_ICONS, NzConfig, NZ_CONFIG, NzMessageService } from 'ng-zorro-antd';
+import { NZ_I18N, zh_CN } from 'ng-zorro-antd/i18n';
+import { NzMessageService } from 'ng-zorro-antd/message';
+import { NZ_CONFIG, NzConfig } from 'ng-zorro-antd/core/config';
 
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
 import { DashboardModule } from './dashboard/dashboard.module';
 import { LoginModule } from './login/login.module';
 import { httpInterceptorProviders } from './interceptor';
-
-import { IconDefinition } from '@ant-design/icons-angular';
-import * as AllIcons from '@ant-design/icons-angular/icons';
 
 // 配置 angular i18n
 import { registerLocaleData } from '@angular/common';
@@ -23,12 +22,8 @@ registerLocaleData(zh);
 
 const ngZorroConfig: NzConfig = {
   message: { nzTop: 20, nzDuration: 3000 },
-  notification: { nzTop: 20, nzDuration: 3000 }
+  notification: { nzTop: 20, nzDuration: 3000 },
 };
-const antDesignIcons = AllIcons as {
-  [key: string]: IconDefinition;
-};
-const icons: IconDefinition[] = Object.keys(antDesignIcons).map(key => antDesignIcons[key]);
 
 @NgModule({
   declarations: [AppComponent, AboutComponent],
@@ -40,16 +35,15 @@ const icons: IconDefinition[] = Object.keys(antDesignIcons).map(key => antDesign
     AppRoutingModule,
     BrowserAnimationsModule,
     ServiceWorkerModule.register('ngsw-worker.js', {
-      enabled: environment.production
-    })
+      enabled: environment.production,
+    }),
   ],
   providers: [
     NzMessageService,
     httpInterceptorProviders,
     { provide: NZ_I18N, useValue: zh_CN },
-    { provide: NZ_ICONS, useValue: icons },
-    { provide: NZ_CONFIG, useValue: ngZorroConfig }
+    { provide: NZ_CONFIG, useValue: ngZorroConfig },
   ],
-  bootstrap: [AppComponent]
+  bootstrap: [AppComponent],
 })
 export class AppModule {}
