@@ -1,7 +1,7 @@
 export const enum Constants {
-  UNLOGIN_FAILED_REDIRECT_URL = 'login',
+  LOGIN_FAILED_REDIRECT_URL = 'login',
   LOGIN_SUCCESS_REDIRECT_URL = 'dashboard',
   LOGIN_USER_NAME = 'user_name',
   USER_TOKEN = 'user_token',
-  LOGIN_URL = '/api/v1/redis/login'
+  LOGIN_URL = '/api/v1/redis/login',
 }
