@@ -3,14 +3,14 @@ import { Router } from '@angular/router';
 import { Constants } from '../constants/constants.enum';
 import { NzMessageService } from 'ng-zorro-antd/message';
 import { DashboardService } from './dashboard.service';
-import { StoreService } from '../store/store.service';
+import { StoreService } from '../services/store.service';
 import { Subject, of } from 'rxjs';
 import { debounceTime, distinctUntilChanged, switchMap } from 'rxjs/operators';
-import { OperaterType, ModalTitle } from '../types/index.interface';
+import { OperateType, ModalTitle } from '../types/index.interface';
 @Component({
   selector: 'app-dashboard',
   templateUrl: './dashboard.component.html',
-  styleUrls: ['./dashboard.component.less']
+  styleUrls: ['./dashboard.component.less'],
 })
 export class DashboardComponent implements OnInit, OnDestroy {
   constructor(
@@ -27,7 +27,7 @@ export class DashboardComponent implements OnInit, OnDestroy {
   backUpKeys: string[];
   searchText$ = new Subject<string>();
   modalTitle: ModalTitle;
-  operatorType: OperaterType;
+  operatorType: OperateType;
 
   @Input() redisTtl = 0;
   @Input() newKey: string;
@@ -43,17 +43,17 @@ export class DashboardComponent implements OnInit, OnDestroy {
       .pipe(
         debounceTime(200),
         distinctUntilChanged(),
-        switchMap(value => {
+        switchMap((value) => {
           let results: string[] = [];
           if (value) {
-            results = this.keys.filter(key => key.includes(value));
+            results = this.keys.filter((key) => key.includes(value));
           } else {
             results = this.backUpKeys;
           }
           return of(results);
         })
       )
-      .subscribe(val => (this.keys = val));
+      .subscribe((val) => (this.keys = val));
     this.handleGetKeys();
     this.handleCurrentUser();
   }
@@ -68,7 +68,7 @@ export class DashboardComponent implements OnInit, OnDestroy {
 
   handleItem(key: string) {
     this.redisKey = key;
-    this.service.handleKeyOfDetail(key).subscribe(val => {
+    this.service.handleKeyOfDetail(key).subscribe((val) => {
       if (val.statusCode === 200) {
         this.content = val.data;
       }
@@ -77,7 +77,7 @@ export class DashboardComponent implements OnInit, OnDestroy {
   }
 
   handleGetKeys() {
-    this.service.handleKeys().subscribe(val => {
+    this.service.handleKeys().subscribe((val) => {
       if (val.statusCode === 200) {
         this.keys = val.data;
         this.backUpKeys = val.data;
@@ -92,13 +92,13 @@ export class DashboardComponent implements OnInit, OnDestroy {
     });
   }
 
-  handleOperator(title: ModalTitle, type: OperaterType) {
+  handleOperator(title: ModalTitle, type: OperateType) {
     this.isVisible = true;
     this.modalTitle = title;
     this.operatorType = type;
   }
 
-  hanldeRefresh(key: string) {
+  handleRefresh(key: string) {
     this.handleKeyOfTtl(key);
   }
 
@@ -106,7 +106,7 @@ export class DashboardComponent implements OnInit, OnDestroy {
     if (!key) {
       return this.message.error('暂无可删除 key');
     }
-    this.service.handleDelete(key).subscribe(val => {
+    this.service.handleDelete(key).subscribe((val) => {
       if (val.statusCode === 200 && val.data) {
         this.handleGetKeys();
         return this.message.success(`${key} 删除成功!`);
@@ -120,7 +120,7 @@ export class DashboardComponent implements OnInit, OnDestroy {
   }
 
   handleRedirectLogin() {
-    this.router.navigateByUrl(Constants.UNLOGIN_FAILED_REDIRECT_URL);
+    this.router.navigateByUrl(Constants.LOGIN_FAILED_REDIRECT_URL);
   }
 
   handleCancel() {
@@ -134,7 +134,7 @@ export class DashboardComponent implements OnInit, OnDestroy {
     this.isVisible = false;
     const { newKey, expireTime, redisValue, operatorType, redisKey } = this;
     if ('add' === operatorType) {
-      return this.service.handleSetKey(newKey, redisValue, expireTime).subscribe(val => {
+      return this.service.handleSetKey(newKey, redisValue, expireTime).subscribe((val) => {
         if (val.statusCode === 200 && val.data) {
           this.handleGetKeys();
           this.newKey = '';
@@ -147,7 +147,7 @@ export class DashboardComponent implements OnInit, OnDestroy {
       });
     }
     if ('resetTtl' === operatorType) {
-      return this.service.handleResetTtlOfKey(redisKey, expireTime).subscribe(val => {
+      return this.service.handleResetTtlOfKey(redisKey, expireTime).subscribe((val) => {
         if (val.statusCode === 200 && val.data) {
           this.handleGetKeys();
           this.expireTime = '';
@@ -158,7 +158,7 @@ export class DashboardComponent implements OnInit, OnDestroy {
       });
     }
     if ('rename' === operatorType) {
-      return this.service.handleRename(redisKey, newKey).subscribe(val => {
+      return this.service.handleRename(redisKey, newKey).subscribe((val) => {
         if (val.statusCode === 200) {
           this.handleGetKeys();
           this.newKey = '';
@@ -171,14 +171,14 @@ export class DashboardComponent implements OnInit, OnDestroy {
   }
 
   handleCurrentUser() {
-    this.service.handleUserInfo().subscribe(val => {
+    this.service.handleUserInfo().subscribe((val) => {
       this.userName = val.data.name;
       this.selectedDb = val.data.db;
     });
   }
 
   private handleKeyOfTtl(key: string) {
-    this.service.handleKeyOfTtl(key).subscribe(val => {
+    this.service.handleKeyOfTtl(key).subscribe((val) => {
       if (val.statusCode === 200) {
         if (val.data === -2) {
           this.redisTtl = 0;
