@@ -1,11 +1,11 @@
 import { Component, OnInit } from '@angular/core';
 import { LoginService } from './login.service';
 import { FormBuilder, FormGroup, Validators } from '@angular/forms';
-import { environment } from '../../environments/environment';
+import { environment } from 'src/environments/environment';
 @Component({
   selector: 'app-login',
   templateUrl: './login.component.html',
-  styleUrls: ['./login.component.less']
+  styleUrls: ['./login.component.less'],
 })
 export class LoginComponent implements OnInit {
   validateForm: FormGroup;
@@ -28,7 +28,7 @@ export class LoginComponent implements OnInit {
       host: ['127.0.0.1', [Validators.required]],
       password: ['1234567890', [Validators.required]],
       port: [6379, [Validators.required]],
-      db: [2, [Validators.required]]
+      db: [2, [Validators.required]],
     };
     if (environment.production) {
       fbObj = {
@@ -36,7 +36,7 @@ export class LoginComponent implements OnInit {
         host: [null, [Validators.required]],
         password: [null, [Validators.required]],
         port: [null, [Validators.required]],
-        db: [null, [Validators.required]]
+        db: [null, [Validators.required]],
       };
     }
     this.validateForm = this.fb.group(fbObj);
