@@ -3,7 +3,10 @@ import { CommonModule } from '@angular/common';
 import { LoginComponent } from './login.component';
 import { BrowserModule } from '@angular/platform-browser';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
-import { NzFormModule, NzInputModule, NzCheckboxModule, NzButtonModule } from 'ng-zorro-antd';
+import { NzFormModule } from 'ng-zorro-antd/form';
+import { NzInputModule } from 'ng-zorro-antd/input';
+import { NzCheckboxModule } from 'ng-zorro-antd/checkbox';
+import { NzButtonModule } from 'ng-zorro-antd/button';
 import { RouterModule } from '@angular/router';
 
 @NgModule({
@@ -17,7 +20,7 @@ import { RouterModule } from '@angular/router';
     NzInputModule,
     NzCheckboxModule,
     NzButtonModule,
-    ReactiveFormsModule
-  ]
+    ReactiveFormsModule,
+  ],
 })
 export class LoginModule {}
