@@ -11,6 +11,6 @@ export interface Response<T> {
   data: T;
 }
 
-export type OperaterType = 'rename' | 'resetTtl' | 'add';
+export type OperateType = 'rename' | 'resetTtl' | 'add';
 
 export type ModalTitle = '重置ttl' | '重命名' | '新增';
